requirements.txt
flask>=3.0.0
openai>=1.3.0
pinecone-client>=3.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-pinecone>=0.0.1
langchain-community>=0.0.10
python-dotenv>=1.0.0
PyPDF2>=3.0.1
docx2txt>=0.8
markdown>=3.5.1
